refactor: extract terminal page hooks
